Fetch remote data into the remote sync buffer table
Refactor synchroniser and split into CentralDataSynchroniser and
RemoteDataSynchroniser
