+++ repository/src/schema/key_value_store.rs
@@ -5,6 +5,11 @@ use diesel_derive_enum::DbEnum;
 #[DbValueStyle = "SCREAMING_SNAKE_CASE"]
 pub enum KeyValueType {
     CentralSyncPullCursor,
+    /// Indicates if the sync queue on the remote server has been initialised
+    RemoteSyncQueueV5Initalised,
+    /// Indicates if the remote data has been pulled and integrated from the central server
+    /// Possible value: "true"
+    RemoteSyncInitialSyncState,
 }
 
 #[derive(Clone, Queryable, Insertable, AsChangeset, Debug, PartialEq, Eq)]
