+++ server/src/sync/central_data_synchroniser.rs
@@ -1,0 +1,289 @@
+use crate::sync::{
+    sync_api_v5::CentralSyncBatchV5,
+    translation::{import_sync_records, SyncImportError, TRANSLATION_RECORDS},
+    SyncApiV5, SyncConnectionError,
+};
+use log::info;
+use repository::{
+    schema::{CentralSyncBufferRow, KeyValueType},
+    CentralSyncBufferRepository, KeyValueStoreRepository, NameStoreJoinRepository, RepositoryError,
+    StorageConnection, TransactionError,
+};
+use thiserror::Error;
+
+use super::sync_api_v5::CentralSyncRecordV5;
+
+#[derive(Error, Debug)]
+pub enum CentralSyncError {
+    #[error("Failed to pull central sync records")]
+    PullCentralSyncRecordsError { source: SyncConnectionError },
+    #[error("Failed to update central sync buffer records")]
+    UpdateCentralSyncBufferRecordsError { source: RepositoryError },
+    #[error("Failed to get central sync cursor record")]
+    GetCentralSyncCursorRecordError { source: RepositoryError },
+    #[error("Failed to get central sync buffer records")]
+    GetCentralSyncBufferRecordsError { source: RepositoryError },
+    #[error("Failed to import central sync buffer records")]
+    ImportCentralSyncRecordsError { source: SyncImportError },
+    #[error("Failed to remove central sync buffer records")]
+    RemoveCentralSyncBufferRecordsError { source: RepositoryError },
+    #[error("Failed to connect to DB")]
+    DBConnectionError { source: RepositoryError },
+}
+
+pub struct CentralDataSynchroniser {
+    pub sync_api_v5: SyncApiV5,
+}
+
+impl CentralDataSynchroniser {
+    async fn pull_central_records(
+        &self,
+        connection: &StorageConnection,
+    ) -> Result<(), CentralSyncError> {
+        let central_sync_cursor = CentralSyncPullCursor::new(&connection);
+        let mut cursor: u32 = central_sync_cursor.get_cursor().unwrap_or_else(|_| {
+            info!("Initialising new central sync cursor...");
+            0
+        });
+
+        // Arbitrary batch size.
+        const BATCH_SIZE: u32 = 500;
+
+        Ok(loop {
+            info!("Pulling {} central sync records...", BATCH_SIZE);
+            let sync_batch: CentralSyncBatchV5 = self
+                .sync_api_v5
+                .get_central_records(cursor, BATCH_SIZE)
+                .await
+                .map_err(|source| CentralSyncError::PullCentralSyncRecordsError { source })?;
+            let central_sync_records = central_sync_batch_records_to_buffer_rows(sync_batch.data)
+                .map_err(|err| {
+                CentralSyncError::PullCentralSyncRecordsError { source: err.into() }
+            })?;
+
+            if central_sync_records.len() == 0 {
+                info!("Central sync buffer is up-to-date");
+                break;
+            }
+
+            info!(
+                "Inserting {} central sync records into central sync buffer",
+                central_sync_records.len()
+            );
+
+            for central_sync_record in central_sync_records {
+                Self::insert_one_and_update_cursor(&connection, &central_sync_record)
+                    .await
+                    .map_err(
+                        |source| CentralSyncError::UpdateCentralSyncBufferRecordsError { source },
+                    )?;
+            }
+            info!("Successfully inserted central sync records into central sync buffer");
+
+            cursor = central_sync_cursor
+                .get_cursor()
+                .map_err(|source| CentralSyncError::GetCentralSyncCursorRecordError { source })?;
+
+            if cursor >= sync_batch.max_cursor - 1 {
+                info!("All central sync records pulled successfully");
+                break;
+            }
+        })
+    }
+
+    /// insert row and update cursor in a single transaction
+    async fn insert_one_and_update_cursor(
+        connection: &StorageConnection,
+        central_sync_buffer_row: &CentralSyncBufferRow,
+    ) -> Result<(), RepositoryError> {
+        let cursor = central_sync_buffer_row.id as u32;
+        // note: if already in a transaction this creates a safepoint:
+        let result: Result<(), TransactionError<RepositoryError>> = connection
+            .transaction(|con| async move {
+                CentralSyncBufferRepository::new(con)
+                    .insert_one(central_sync_buffer_row)
+                    .await?;
+                CentralSyncPullCursor::new(con).update_cursor(cursor)?;
+                Ok(())
+            })
+            .await;
+        Ok(result?)
+    }
+
+    async fn integrate_central_records(
+        &self,
+        connection: &StorageConnection,
+    ) -> Result<(), CentralSyncError> {
+        let central_sync_buffer_repository = CentralSyncBufferRepository::new(&connection);
+
+        let mut records: Vec<CentralSyncBufferRow> = Vec::new();
+        for table_name in TRANSLATION_RECORDS {
+            info!("Querying central sync buffer for {} records", table_name);
+
+            let mut buffer_rows = central_sync_buffer_repository
+                .get_sync_entries(table_name)
+                .await
+                .map_err(|source| CentralSyncError::GetCentralSyncBufferRecordsError { source })?;
+
+            info!(
+                "Found {} {} records in central sync buffer",
+                buffer_rows.len(),
+                table_name
+            );
+
+            records.append(&mut buffer_rows);
+        }
+
+        info!("Importing {} central sync buffer records...", records.len());
+        import_sync_records(connection, &records)
+            .await
+            .map_err(|source| CentralSyncError::ImportCentralSyncRecordsError { source })?;
+        info!("Successfully Imported central sync buffer records",);
+
+        // TODO needs to be done for M1 as name_store_joins are not synced yet but are required in API
+        // these records should actually sync from server in remote sync
+        if records.len() > 0 {
+            match NameStoreJoinRepository::new(connection).m1_add() {
+                Ok(_) => {}
+                Err(_) => {}
+            };
+        }
+
+        info!("Clearing central sync buffer");
+        central_sync_buffer_repository
+            .remove_all()
+            .await
+            .map_err(|source| CentralSyncError::RemoveCentralSyncBufferRecordsError { source })?;
+        info!("Successfully cleared central sync buffer");
+
+        Ok(())
+    }
+
+    pub async fn pull(&self, connection: &StorageConnection) -> Result<(), CentralSyncError> {
+        info!("Syncing central records...");
+        self.pull_central_records(connection).await?;
+        info!("Successfully synced central records");
+
+        info!("Integrating central records...");
+        self.integrate_central_records(connection).await?;
+        info!("Successfully integrated central records");
+
+        // info!("Syncing remote records...");
+        // let remote_records = self.pull_remote_records().await?;
+        // info!("Successfully pulled remote records");
+
+        // info!("Integrating remote records...");
+        // self.integrate_remote_records(remote_records);
+        // info!("Successfully integrated remote records");
+
+        Ok(())
+    }
+}
+
+fn central_sync_batch_records_to_buffer_rows(
+    records: Option<Vec<CentralSyncRecordV5>>,
+) -> Result<Vec<CentralSyncBufferRow>, serde_json::Error> {
+    let central_sync_records: Result<Vec<CentralSyncBufferRow>, serde_json::Error> = records
+        .unwrap_or(vec![])
+        .into_iter()
+        .map(|record| {
+            Ok(CentralSyncBufferRow {
+                id: record.id,
+                table_name: record.table_name,
+                record_id: record.record_id,
+                data: serde_json::to_string(&record.data)?,
+            })
+        })
+        .collect();
+    central_sync_records
+}
+
+struct CentralSyncPullCursor<'a> {
+    key_value_store: KeyValueStoreRepository<'a>,
+}
+
+impl<'a> CentralSyncPullCursor<'a> {
+    pub fn new(connection: &'a StorageConnection) -> Self {
+        CentralSyncPullCursor {
+            key_value_store: KeyValueStoreRepository::new(connection),
+        }
+    }
+
+    pub fn get_cursor(&self) -> Result<u32, RepositoryError> {
+        let value = self
+            .key_value_store
+            .get_string(KeyValueType::CentralSyncPullCursor)?;
+        let cursor = value
+            .and_then(|value| value.parse::<u32>().ok())
+            .unwrap_or(0);
+        Ok(cursor)
+    }
+
+    pub fn update_cursor(&self, cursor: u32) -> Result<(), RepositoryError> {
+        self.key_value_store.set_string(
+            KeyValueType::CentralSyncPullCursor,
+            Some(format!("{}", cursor)),
+        )
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        sync::{
+            translation::test_data::{
+                check_records_against_database, extract_sync_buffer_rows,
+                item::get_test_item_records, master_list::get_test_master_list_records,
+                master_list_line::get_test_master_list_line_records,
+                master_list_name_join::get_test_master_list_name_join_records,
+                name::get_test_name_records, store::get_test_store_records,
+            },
+            SyncApiV5, SyncCredentials,
+        },
+        test_utils::get_test_settings,
+    };
+    use repository::{
+        get_storage_connection_manager, schema::CentralSyncBufferRow, test_db,
+        CentralSyncBufferRepository,
+    };
+    use reqwest::{Client, Url};
+
+    use super::CentralDataSynchroniser;
+
+    #[actix_rt::test]
+    async fn test_integrate_central_records() {
+        let settings = get_test_settings("omsupply-database-integrate_central_records");
+
+        test_db::setup(&settings.database).await;
+        let connection_manager = get_storage_connection_manager(&settings.database);
+
+        // use test records with cursors that are out of order
+        let mut test_records = Vec::new();
+        test_records.append(&mut get_test_name_records());
+        test_records.append(&mut get_test_item_records());
+        test_records.append(&mut get_test_store_records());
+        test_records.append(&mut get_test_master_list_records());
+        test_records.append(&mut get_test_master_list_name_join_records());
+        test_records.append(&mut get_test_master_list_line_records());
+
+        let central_records: Vec<CentralSyncBufferRow> = extract_sync_buffer_rows(&test_records);
+        let connection = connection_manager.connection().unwrap();
+        let central_sync_buffer_repository = CentralSyncBufferRepository::new(&connection);
+
+        central_sync_buffer_repository
+            .insert_many(&central_records)
+            .expect("Failed to insert central sync records into sync buffer");
+
+        let client = Client::new();
+        let url = Url::parse(&settings.sync.url).unwrap();
+        let credentials = SyncCredentials::new(&settings.sync.username, &settings.sync.password);
+        let sync_api_v5 = SyncApiV5::new(url, credentials, client);
+
+        let sync = CentralDataSynchroniser { sync_api_v5 };
+        sync.integrate_central_records(&connection)
+            .await
+            .expect("Failed to integrate central records");
+
+        check_records_against_database(&connection, test_records).await;
+    }
+}
