+++ server/src/sync/mod.rs
@@ -1,4 +1,6 @@
 mod actor;
+mod central_data_synchroniser;
+mod remote_data_synchroniser;
 mod sync_api_credentials;
 mod sync_api_v3;
 mod sync_api_v5;
@@ -8,4 +10,4 @@ mod translation;
 pub use actor::{get_sync_actors, SyncReceiverActor, SyncSenderActor};
 pub use sync_api_credentials::SyncCredentials;
 pub use sync_api_v5::{SyncApiV5, SyncConnectionError};
-pub use synchroniser::{CentralSyncError, RemoteSyncError, SyncError, Synchroniser};
+pub use synchroniser::Synchroniser;
